requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
faiss-cpu==1.13.0
filelock==3.19.1
fsspec==2025.10.0
gensim==4.4.0
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
smart_open==7.5.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wrapt==2.1.1
